requirements.txt
# Core extras kept slim for fast startup; add more providers as needed.
apache-airflow==2.9.2
apache-airflow-providers-http==4.12.0
apache-airflow-providers-common-io==1.4.0
httpx==0.27.2
beautifulsoup4==4.12.3
duckdb==1.1.3
pandas==2.2.2
pydantic==2.9.2
python-dateutil==2.9.0.post0
